refactor: improved CartDrawer structure
