+++ react-client/src/components/Membership/CartDrawer.tsx
@@ -1,6 +1,7 @@
-import { useEffect, useState } from "react";
-import { formatPrice, getValueFromPrice } from "~/util/stringUtil";
-import { Button, Drawer, Icon, Text } from "../design-system";
+import { useCartItems } from "~/hooks/useCartItems";
+import { Drawer, Text } from "../design-system";
+import { CartFooter } from "./CartFooter";
+import { CartHeader } from "./CartHeader";
 import { CartItemCard } from "./CartItemCard";
 import { CartDrawerProps } from "./utils/cartTypes";
 
@@ -11,37 +12,17 @@ export const CartDrawer = ({
 }: CartDrawerProps & {
   updateQuantity: (id: string, value: number) => void;
 }) => {
-  const [subtotal, setSubtotal] = useState("");
-
-  useEffect(() => {
-    function calculateSubtotal(): number {
-      return items.reduce((total, item) => {
-        const itemPrice = getValueFromPrice(item.price);
-        const itemQuantity = item.quantity ?? 1;
-        return total + itemPrice * itemQuantity;
-      }, 0);
-    }
-
-    setSubtotal(formatPrice(calculateSubtotal()));
-  }, [items]);
+  const { subtotal } = useCartItems(items);
 
   return (
     <Drawer id="cart-drawer" ariaLabel="Cart Drawer" {...props}>
       <div className="flex flex-col gap-large">
-        <div className="flex items-center">
-          <Icon
-            display="shoppingCart"
-            className="mr-small text-orange-300-contrast"
-          />
-          <Text fontWeight="bold" size="xlg">
-            My Cart
-          </Text>
-        </div>
-
+        <CartHeader />
         <div className="flex flex-col gap-small">
           <Text size="base" fontWeight="bold">
             Items
           </Text>
+
           {items.map((props) => (
             <CartItemCard
               key={props.id}
@@ -51,21 +32,7 @@ export const CartDrawer = ({
           ))}
         </div>
 
-        <div className="flex w-full justify-between">
-          <div className="w-1/2">
-            <Text fontWeight="bold" display="block">
-              Subtotal
-            </Text>
-            <Text color="background-color-tertiary">
-              Applicable taxes will be applied at checkout
-            </Text>
-          </div>
-          <Text fontWeight="bold" size="xlg">
-            {`$${subtotal}`}
-          </Text>
-        </div>
-
-        <Button>Proceed to checkout</Button>
+        <CartFooter subtotal={subtotal} />
       </div>
     </Drawer>
   );
